+++ cmd/mongodb/fnQueryData.go
@@ -22,3 +22,68 @@ func GetAllResources(coll *mongo.Collection) (documentData []lib.AzureResourceDe
 
 	return documentData
 }
+
+
+
+//
+//
+
+func GetResourceSKUs(collection *mongo.Collection) (skus []lib.AzureResourceSku) {
+	ctx := context.Background()
+
+	filter := bson.D{{}}
+
+	rsp, err := collection.Find(ctx, filter)
+	lib.CheckFatalError(err)
+
+	err = rsp.All(ctx, &skus)
+	lib.CheckFatalError(err)
+	// if len(skus) == 0 {
+	// 	fmt.Println("No apps in slice")
+	// 	return nil
+	// }
+	// ctx := context.TODO()
+
+	// var updates []mongo.WriteModel
+
+	// for _, sku := range skus {
+	// 	curr := sku
+	// 	curr.LastDBSync = time.Now()
+	// 	filter := bson.D{{"_id", sku.Name}}
+	// 	update := bson.D{{"$set", curr}}
+
+	// 	// .SetUpsert(true)
+	// 	updates = append(updates, mongo.NewUpdateOneModel().SetFilter(filter).SetUpdate(update).SetUpsert(true))
+	// }
+
+	// var opts options.BulkWriteOptions
+	// opts.SetOrdered(false)
+
+	// chunkSize := 100
+	// var chunks [][]mongo.WriteModel
+	// for i := 0; i < len(updates); i += chunkSize {
+	// 	end := i + chunkSize
+	// 	if end > len(updates) {
+	// 		end = len(updates)
+	// 	}
+	// 	chunks = append(chunks, updates[i:end])
+	// }
+
+	// for _, chunk := range chunks {
+	// 	res, err := collection.BulkWrite(ctx, chunk, &opts)
+	// 	results = append(results, *res)
+	// 	lib.CheckFatalError(err)
+	// }
+
+	// // fmt.Printf("Number of documents inserted: %d\n", results.InsertedCount)
+	// // fmt.Printf("Number of documents matched: %d\n", results.MatchedCount)
+	// // fmt.Printf("Number of documents matched: %d\n", )
+	// // fmt.Printf("Number of documents inserted: %d\n", results.UpsertedCount)
+	// // fmt.Printf("Number of documents replaced or updated: %d\n", results.ModifiedCount)
+	// // fmt.Printf("Number of documents deleted: %d\n", results.DeletedCount)
+	// // fmt.Println("Upserted IDs:")
+	// // jsonStr, _ := json.MarshalIndent(results.UpsertedIDs, "", "  ")
+	// // fmt.Println(string(jsonStr))
+	// return results
+	return
+}